Split UserInterface into new DBCommunicator.
